refactor(우승자 찾기 책임 분리)
