+++ src/main/kotlin/racingcar/RacingCarGame.kt
@@ -2,19 +2,15 @@ package racingcar
 
 class RacingCarGame {
 
-    fun progressGame(totalRound: Long, cars: List<Car>): RacingCarGameResult {
+    fun progressGame(totalRound: Long, cars: List<Car>): List<RacingCarRoundResult> {
         require(totalRound >= 1) { "시도 횟수는 1 이상이어야 합니다." }
         require(cars.isNotEmpty()) { "자동차 대수는 1 이상이어야 합니다." }
 
-        val gameRoundResults = (START_ROUND..totalRound)
+        return (START_ROUND..totalRound)
             .map { round ->
                 val roundResult = progressRound(cars)
                 RacingCarRoundResult(round, roundResult)
             }
-
-        val winnerNames = findWinnerNames(gameRoundResults = gameRoundResults, lastRound = totalRound)
-
-        return RacingCarGameResult(winnerNames, gameRoundResults)
     }
 
     private fun progressRound(cars: List<Car>): List<RacingCarDriveResult> {
@@ -23,26 +19,6 @@ class RacingCarGame {
             RacingCarDriveResult(car.carNumber, car.position, car.name)
         }
     }
-
-    private fun findWinnerNames(gameRoundResults: List<RacingCarRoundResult>, lastRound: Long): List<String> {
-        val lastRoundResults = getLastRoundResults(gameRoundResults, lastRound)
-
-        val maxPositionCarResult = lastRoundResults.maxBy { it.position }
-
-        return lastRoundResults.asSequence()
-            .filter { it.position == maxPositionCarResult.position }
-            .map { it.carName }
-            .toList()
-    }
-
-    private fun getLastRoundResults(
-        gameRoundResults: List<RacingCarRoundResult>,
-        lastRound: Long
-    ) = gameRoundResults.asSequence()
-        .filter { it.round == lastRound }
-        .map { it.carDriveResults }
-        .flatten()
-        .toList()
 
     companion object {
         const val MIN_RANDOM_DRIVE_NUMBER: Long = 4
